[feat]: Moved cli commands common to testing, cli, and sdk out into apps
