+++ apps/src/bin/namada-relayer/main.rs
@@ -1,7 +1,7 @@
-mod cli;
-
 use color_eyre::eyre::Result;
-use namada_apps::logging;
+use namada::tendermint_rpc::HttpClient;
+use namada_apps::cli::api::CliApi;
+use namada_apps::{cli, logging};
 use tracing_subscriber::filter::LevelFilter;
 
 #[tokio::main]
@@ -12,6 +12,7 @@ async fn main() -> Result<()> {
     // init logging
     logging::init_from_env_or(LevelFilter::INFO)?;
 
+    let (cmd, _) = cli::namada_relayer_cli()?;
     // run the CLI
-    cli::main().await
+    CliApi::<()>::handle_relayer_command::<HttpClient>(None, cmd).await
 }
